backend/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
python-multipart==0.0.9
pydantic==2.6.3

# ML model
torch==2.1.0
transformers==4.39.3
Pillow==10.2.0

# Groq API client
groq==0.9.0

# ElevenLabs TTS
elevenlabs==1.3.1

# Utilities
python-dotenv==1.0.1
requests==2.31.0
